requirements-vercel.txt
Django==5.0.7
gunicorn==22.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
django-tinymce==3.7.1
python-decouple==3.8
django-cors-headers==4.3.1
django-storages==1.14.2
